% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{beta}
\alias{beta}
\title{Beta random vectors}
\usage{
beta(a, b)
}
\arguments{
\item{a}{Number of successes \eqn{\alpha}}

\item{b}{Number of failures \eqn{\beta}}
}
\value{
A Beta random vector
}
\description{
Generate and do operations on Beta random vectors
}
\examples{
x <- beta(1:10, 10:1)
x
mean(x)
draw(x)

}

