% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{draw}
\alias{draw}
\title{Random Vectors
Draw from a random vector}
\usage{
draw(rv)
}
\arguments{
\item{rv}{A random vector}
}
\value{
A numeric vector of same length
}
\description{
Random Vectors
Draw from a random vector
}

