% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{gamma}
\alias{gamma}
\title{Gamma random vectors}
\usage{
gamma(a, b)
}
\arguments{
\item{a}{Number of events \eqn{\alpha}}

\item{b}{Time elapsed \eqn{\beta}}
}
\value{
A Gamma random vector
}
\description{
Generate and do operations on Gamma random vectors
}
\examples{
x <- gamma(1:10, 10)
x
mean(x)
draw(x)

}
