% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli.R
\name{infer_bernoulli}
\alias{infer_bernoulli}
\title{Infer Bernoulli process}
\usage{
infer_bernoulli(obs)
}
\arguments{
\item{obs}{A series of Bernoulli trials}
}
\value{
a Beta random variable for the probability \eqn{p}
}
\description{
Infer a Bernoulli process from data
}
