% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{infer_poisson}
\alias{infer_poisson}
\title{Infer Poisson process}
\usage{
infer_poisson(obs)
}
\arguments{
\item{obs}{A series of Poisson trials}
}
\value{
a Gamma random variable for the rate \eqn{\lambda}
}
\description{
Infer a Poisson process from data
}
