% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnifyr.R
\docType{package}
\name{magnifyr}
\alias{magnifyr}
\alias{magnifyr-package}
\title{Magnifyr: magnify your inferences using random objects}
\description{
Implements random vectors and related methods including inference and
simulation. Designed to work smoothly with data frames.
}
\section{Distributions}{

\code{\link{beta}}
}

