% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{normal}
\alias{normal}
\title{Normal random vectors}
\usage{
normal(mu, sigma)
}
\arguments{
\item{mu}{Mean \eqn{\mu}}

\item{sigma}{Standard deviation \eqn{\sigma}}
}
\value{
A Normal random vector
}
\description{
Generate and do operations on Normal random vectors
}
\examples{
x <- normal(1:10, 10)
x
mean(x)
draw(x)

}
